requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
click==8.1.3
Flask==2.2.2
Flask-WTF==1.1.1
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.2
zipp==3.11.0
lxml==4.9.2

pip~=22.3.1
setuptools~=58.0.4
mysql-connector-python~=8.0.32